+++ src/siteSettings/components/SiteSettingsMailing/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./SiteSettingsMailing";
+export * from "./SiteSettingsMailing";