+++ src/siteSettings/components/SiteSettingsPage/SiteSettingsPage.tsx
@@ -1,4 +1,6 @@
+import { Theme } from "@material-ui/core/styles";
 import Typography from "@material-ui/core/Typography";
+import { makeStyles } from "@material-ui/styles";
 import React from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
@@ -7,6 +9,7 @@ import { ConfirmButtonTransitionState } from "@saleor/components/ConfirmButton";
 import Container from "@saleor/components/Container";
 import Form from "@saleor/components/Form";
 import Grid from "@saleor/components/Grid";
+import Hr from "@saleor/components/Hr";
 import PageHeader from "@saleor/components/PageHeader";
 import SaveButtonBar from "@saleor/components/SaveButtonBar";
 import useStateFromProps from "@saleor/hooks/useStateFromProps";
@@ -20,6 +23,9 @@ import { SiteSettings_shop } from "../../types/SiteSettings";
 import SiteSettingsAddress from "../SiteSettingsAddress/SiteSettingsAddress";
 import SiteSettingsDetails from "../SiteSettingsDetails/SiteSettingsDetails";
 import SiteSettingsKeys from "../SiteSettingsKeys/SiteSettingsKeys";
+import SiteSettingsMailing, {
+  SiteSettingsMailingFormData
+} from "../SiteSettingsMailing";
 
 export interface SiteSettingsPageAddressFormData {
   city: string;
@@ -33,7 +39,8 @@ export interface SiteSettingsPageAddressFormData {
 }
 
 export interface SiteSettingsPageFormData
-  extends SiteSettingsPageAddressFormData {
+  extends SiteSettingsPageAddressFormData,
+    SiteSettingsMailingFormData {
   description: string;
   domain: string;
   name: string;
@@ -50,33 +57,53 @@ export interface SiteSettingsPageProps {
   onSubmit: (data: SiteSettingsPageFormData) => void;
 }
 
-const SiteSettingsPage: React.StatelessComponent<SiteSettingsPageProps> = ({
-  disabled,
-  errors,
-  saveButtonBarState,
-  shop,
-  onBack,
-  onKeyAdd,
-  onKeyRemove,
-  onSubmit
-}) => {
+const useStyles = makeStyles(
+  (theme: Theme) => ({
+    hr: {
+      gridColumnEnd: "span 2",
+      margin: `${theme.spacing.unit}px 0`
+    }
+  }),
+  {
+    name: "SiteSettingsPage"
+  }
+);
+
+const SiteSettingsPage: React.FC<SiteSettingsPageProps> = props => {
+  const {
+    disabled,
+    errors,
+    saveButtonBarState,
+    shop,
+    onBack,
+    onKeyAdd,
+    onKeyRemove,
+    onSubmit
+  } = props;
+  const classes = useStyles(props);
   const intl = useIntl();
   const [displayCountry, setDisplayCountry] = useStateFromProps(
     maybe(() => shop.companyAddress.country.code, "")
   );
 
-  const initialForm: SiteSettingsPageFormData = {
+  const initialFormAddress: SiteSettingsPageAddressFormData = {
     city: maybe(() => shop.companyAddress.city, ""),
     companyName: maybe(() => shop.companyAddress.companyName, ""),
     country: maybe(() => shop.companyAddress.country.code, ""),
     countryArea: maybe(() => shop.companyAddress.countryArea, ""),
-    description: maybe(() => shop.description, ""),
-    domain: maybe(() => shop.domain.host, ""),
-    name: maybe(() => shop.name, ""),
     phone: maybe(() => shop.companyAddress.phone, ""),
     postalCode: maybe(() => shop.companyAddress.postalCode, ""),
     streetAddress1: maybe(() => shop.companyAddress.streetAddress1, ""),
     streetAddress2: maybe(() => shop.companyAddress.streetAddress2, "")
+  };
+  const initialForm: SiteSettingsPageFormData = {
+    ...initialFormAddress,
+    customerSetPasswordUrl: maybe(() => shop.customerSetPasswordUrl, ""),
+    defaultMailSenderAddress: maybe(() => shop.defaultMailSenderAddress, ""),
+    defaultMailSenderName: maybe(() => shop.defaultMailSenderName, ""),
+    description: maybe(() => shop.description, ""),
+    domain: maybe(() => shop.domain.host, ""),
+    name: maybe(() => shop.name, "")
   };
 
   return (
@@ -105,18 +132,51 @@ const SiteSettingsPage: React.StatelessComponent<SiteSettingsPageProps> = ({
               title={intl.formatMessage(commonMessages.generalInformations)}
             />
             <Grid variant="inverted">
-              <Typography variant="h6">
-                {intl.formatMessage(sectionNames.siteSettings)}
-              </Typography>
+              <div>
+                <Typography>
+                  {intl.formatMessage(sectionNames.siteSettings)}
+                </Typography>
+                <Typography variant="body1">
+                  <FormattedMessage defaultMessage="These are general information about your store. They define what is the URL of your store and what is shown in browsers taskbar." />
+                </Typography>
+              </div>
               <SiteSettingsDetails
                 data={data}
                 errors={formErrors}
                 disabled={disabled}
                 onChange={change}
               />
-              <Typography variant="h6">
-                <FormattedMessage defaultMessage="Company information" />
-              </Typography>
+              <Hr className={classes.hr} />
+              <div>
+                <Typography>
+                  <FormattedMessage
+                    defaultMessage="Mailing Configuration"
+                    description="section header"
+                  />
+                </Typography>
+                <Typography variant="body1">
+                  <FormattedMessage defaultMessage="This where you will find all of the settings determining your stores e-mails. You can determine main email address and some of the contents of your emails." />
+                </Typography>
+              </div>
+              <SiteSettingsMailing
+                data={data}
+                errors={formErrors}
+                disabled={disabled}
+                onChange={change}
+              />
+              <Hr className={classes.hr} />
+              <div>
+                <Typography>
+                  <FormattedMessage
+                    defaultMessage="Company Information"
+                    description="section header"
+                  />
+                </Typography>
+                <Typography variant="body1">
+                  <FormattedMessage defaultMessage="This adress will be used to generate invoices and calculate shipping rates." />
+                  <FormattedMessage defaultMessage="Email adress you provide here will be used as a contact adress for your customers." />
+                </Typography>
+              </div>
               <SiteSettingsAddress
                 data={data}
                 displayCountry={displayCountry}
@@ -126,9 +186,18 @@ const SiteSettingsPage: React.StatelessComponent<SiteSettingsPageProps> = ({
                 onChange={change}
                 onCountryChange={handleCountryChange}
               />
-              <Typography variant="h6">
-                <FormattedMessage defaultMessage="Authentication keys" />
-              </Typography>
+              <Hr className={classes.hr} />
+              <div>
+                <Typography>
+                  <FormattedMessage
+                    defaultMessage="Authentication Methods"
+                    description="section header"
+                  />
+                </Typography>
+                <Typography variant="body1">
+                  <FormattedMessage defaultMessage="Authentication method defines additional ways that customers can log in to your ecommerce. " />
+                </Typography>
+              </div>
               <SiteSettingsKeys
                 disabled={disabled}
                 keys={maybe(() => shop.authorizationKeys)}
@@ -148,5 +217,6 @@ const SiteSettingsPage: React.StatelessComponent<SiteSettingsPageProps> = ({
     </Form>
   );
 };
+
 SiteSettingsPage.displayName = "SiteSettingsPage";
 export default SiteSettingsPage;