+++ CHANGELOG.md
@@ -46,4 +46,5 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix attribute errors - #216 by @dominik-zeglen
 - Fix column picker - #228 by @dominik-zeglen
 - Add readonly mode - #229 by @dominik-zeglen
+- Add mailing configuration - #222 by @dominik-zeglen
 - Fix minor bugs - #230 by @dominik-zeglen